Base完成

+++ parts.py
@@ -2,6 +2,7 @@ from datetime import datetime
 from db_config import Day
 from db_config import Number
 from dateutil import relativedelta
+import statistics
 
 
 """ リスト作成 """
@@ -23,7 +24,6 @@ def list_number(pig_no):  # numberリスト
         f"{number.born_num10}",
         f"{number.born_num11}",
         f"{number.born_num12}",
-        # f"{number.delete_day}",
         f"{number.pub_datetime}",
     ]
 
@@ -95,13 +95,12 @@ def create(new_no, new_day):
         born_num10=0,
         born_num11=0,
         born_num12=0,
-        # delete_day="3000/1/1",
     )
 
 
 # 新data登録コード
 # new_no = "77-77"
-# new_day = "2020/7/7"
+# new_day = "2019/7/7"
 # create(new_no, new_day)
 
 """ Rコマンドで使用 """
@@ -217,7 +216,7 @@ def in_born_data12(pig_no, new_day, new_num):  # born12用関数
 
 # 出産情報登録コード
 # pig_no = "28-10"
-# new_day = 2022/7/7
+# new_day = input("出産日は？")
 # new_num = 10
 
 # in_born_data1(pig_no, new_day, new_num)  # born1入力
@@ -240,8 +239,7 @@ def in_born_data12(pig_no, new_day, new_num):  # born12用関数
 def pig_info(pig_no, number_l, day_l):
     born_span_l = []
     for n in range(2, 13):
-        born_span = datetime.strptime(
-            day_l[n + 1], "%Y/%m/%d") - datetime.strptime(
+        born_span = datetime.strptime(day_l[n + 1], "%Y/%m/%d") - datetime.strptime(
             day_l[n], "%Y/%m/%d"
         )
         born_span_l.append(born_span)
@@ -279,86 +277,37 @@ def pig_info(pig_no, number_l, day_l):
 # day_l = list_day(pig_no)
 # pig_info(pig_no, number_l, day_l)
 
-# 年齢
-# pig_no = "28-10"
-# p = Day.get(Day.pig_no == pig_no)
-# print(f"{p.add_day}")
-# print(type(p.add_day))
-# ad = datetime.strptime(p.add_day, "%Y/%m/%d")
-# print(ad)
-# print(type(ad))
-# str_ad = ad.strftime("%Y%m%d")
-# print(str_ad)
-# print(type(str_ad))
 
-# today = datetime.now()
-# print(today)
-# print(type(today))
-# str_td = today.strftime("%Y%m%d")
-# print(str_td)
-# print(type(str_td))
-
-# old = int((int(str_td) - int(str_ad) + 600)/10000)
-# print(old)
+""" Cコマンドで使用 """
 
 
-# index 3...day4-day3
-# 回転数はday3-day2...[idx-1]
-# 出産日はday4...List[5]...[idx+2]
-# 出産頭数はnum4...List[4]...[idx+1]
-# tdatetime = datetime.strptime(tstr, "%Y-%m-%d %H:%M:%S")
-
-# 以下をインライン化して関数
-# time_l = []
-# for n in range(2, 13):
-# t_day = datetime.strptime(List[n], "%Y-%m-%d")
-# time_l.append(t_day)
-
-# print(time_l)
-# day_timne_l = []
-# for n in range(10):
-# day_time = time_l[n+1] - time_l[n]
-# day_timne_l.append(day_time)
+def delete_day_delete():  # delete_dayの日付がtodayより前のデータを削除
+    for p in Day.select():
+        pig_no = p.pig_no
+        day = Day.get(Day.pig_no == pig_no)
+        # print(type(day.delete_day))
+        delete_day = datetime.strptime(day.delete_day, "%Y/%m/%d").date()
+        today = datetime.now().date()  # .date()でdate型に変換
+        if (delete_day - today).days < 0:
+            number = Number.get(Number.pig_no == pig_no)
+            day = Day.get(Day.pig_no == pig_no)
+            number.delete_instance()
+            day.delete_instance()
+        else:
+            pass
 
 
-""" C,Dコマンドで使用 """
-
-# Cコマンドでdelete_dayの日付がtodayより前のデータは削除
-# 回転数0以上の個体(まだ1回しか出産していない個体)について、平均算出
+# delete_day_delete()
 
 
-def delete(pig_no):  # 単純削除機能
-    number = Number.get(Number.pig_no == pig_no)
-    day = Day.get(Day.pig_no == pig_no)
-    number.delete_instance()
-    day.delete_instance()
-
-
-# pig_no = "99-99"
-# delete(pig_no)
-
-
-""" Sコマンドで使用 """
-
-
-def search(list_number, list_day):
+def check(list_day):  # 2回以上出産した個体の平均算出
     all_pig_l = []  # 空リスト用意(すべてのpigの辞書が入る)
     for p in Day.select():
         pig_no = p.pig_no
-        d = Day.get(Day.pig_no == pig_no)
-
-        #  age算出
-        add_day = (datetime.strptime(d.add_day, "%Y/%m/%d")).strftime("%Y%m%d")
-        today = (datetime.now()).strftime("%Y%m%d")
-        age = int((int(today) - int(add_day) + 600) / 10000)
-
-        # rotate算出
-        number_l = list_number(pig_no)
         day_l = list_day(pig_no)
         born_span_l = []
         for n in range(2, 13):
-            born_span = datetime.strptime(
-                day_l[n + 1], "%Y/%m/%d") - datetime.strptime(
+            born_span = datetime.strptime(day_l[n + 1], "%Y/%m/%d") - datetime.strptime(
                 day_l[n], "%Y/%m/%d"
             )
             born_span_l.append(born_span)
@@ -383,9 +332,74 @@ def search(list_number, list_day):
             else:
                 pass
 
-        span = (
-            datetime.now() - datetime.strptime(day_l[idx + 2], "%Y/%m/%d")
-            ).days
+        # 辞書リスト作成
+        key = ["pig_no", "rotate"]
+        value = [pig_no, f"{round(rotate,2)}"]
+        p_dic = dict(zip(key, value))  # すべてのpigの辞書作成
+        all_pig_l.append(p_dic)  # すべてのpigの辞書をリスト化
+    r_l = []
+    for i in range(len(all_pig_l)):  # 1つずつpigの辞書を取り出す
+        p_info = all_pig_l[i]
+        r = float(p_info["rotate"])
+        if 0 < r:
+            r_l.append(r)
+        else:
+            pass
+    print(r_l)
+
+    mean = statistics.mean(r_l)
+    print(f"平均回転数: {round(mean, 2)}回")
+
+
+# 全体の平均回転数を算出
+# check(list_day)
+
+
+""" Sコマンドで使用 """
+
+
+def search(list_number, list_day):
+    all_pig_l = []  # 空リスト用意(すべてのpigの辞書が入る)
+    for p in Day.select():
+        pig_no = p.pig_no
+        d = Day.get(Day.pig_no == pig_no)
+
+        #  age算出
+        add_day = (datetime.strptime(d.add_day, "%Y/%m/%d")).strftime("%Y%m%d")
+        today = (datetime.now()).strftime("%Y%m%d")
+        age = int((int(today) - int(add_day) + 600) / 10000)
+
+        # rotate算出
+        number_l = list_number(pig_no)
+        day_l = list_day(pig_no)
+        born_span_l = []
+        for n in range(2, 13):
+            born_span = datetime.strptime(day_l[n + 1], "%Y/%m/%d") - datetime.strptime(
+                day_l[n], "%Y/%m/%d"
+            )
+            born_span_l.append(born_span)
+        idx = 10  # 初期値を設定 possibly unbound 回避
+        print(born_span_l)
+        for n in range(0, 10):  # 直近の出産日index取得
+            if born_span_l[n].days < 0:
+                idx = born_span_l.index(born_span_l[n])
+                if idx == 0:  # pig_noが産子数(2)に入る防止
+                    idx = 1
+                else:
+                    pass
+                break
+            else:
+                idx = 10
+        if born_span_l[(idx - 1)].days == 0:
+            rotate = 0  # division by zero 回避
+        else:
+            rotate = 365 / born_span_l[(idx - 1)].days  # 回転数算出
+            if rotate < 0:  # 過去1回しか出産していない場合rotateがマイナスになる防止
+                rotate = 0
+            else:
+                pass
+
+        span = (datetime.now() - datetime.strptime(day_l[idx + 2], "%Y/%m/%d")).days
 
         # 辞書リスト作成
         key = ["pig_no", "age", "rotate", "num1", "num2", "span"]
@@ -418,62 +432,26 @@ def search(list_number, list_day):
         else:
             pass
 
+
+# 生産性低い個体を探すキー
 # search(list_number, list_day)
-
-
-"""
-# age算出
-pig_no = "29-37"
-d = Day.get(Day.pig_no == pig_no)
-add_day = (datetime.strptime(d.add_day, "%Y/%m/%d")).strftime("%Y%m%d")
-today = (datetime.now()).strftime("%Y%m%d")
-age = int((int(today) - int(add_day) + 600) / 10000)
-
-# rotate算出
-number_l = list_number(pig_no)
-day_l = list_day(pig_no)
-day_time_l = []
-for n in range(2, 13):
-    t_day = datetime.strptime(day_l[n + 1], "%Y/%m/%d") - datetime.strptime(
-        day_l[n], "%Y/%m/%d"
-    )
-    day_time_l.append(t_day)
-idx = 10  # 初期値を設定 possibly unbound 回避
-for n in range(0, 10):  # 直近の出産日index取得
-    if day_time_l[n].days < 0:
-        idx = day_time_l.index(day_time_l[n])
-        break
-    else:
-        idx = 10
-if day_time_l[(idx - 1)].days == 0:
-    rotate = 0  # division by zero 回避
-else:
-    rotate = 365 / day_time_l[(idx - 1)].days  # 回転数算出
-
-
-key = ["pig_no", "age", "rotate", "num1", "num2"]
-value = [pig_no, age, f"{round(rotate,2)}", number_l[idx], number_l[idx + 1]]
-p_dic = dict(zip(key, value))
-print(p_dic)
-print(type(value[3]))
-"""
 
 
 """ Dコマンドで使用 """
 
 
-def delete_day_set(pig_no):  # 時限爆弾設置
-    # number = Number.get(Number.pig_no == pig_no)
+def delete_day_set(pig_no):  # 時限爆弾設置(str型でDBに登録)
     day = Day.get(Day.pig_no == pig_no)
-    delete_day = datetime.now() + relativedelta.relativedelta(years=1)
-    # number.delete_day = delete_day
-    # number.save()
+    delete_day = (datetime.now() + relativedelta.relativedelta(years=1)).strftime(
+        "%Y/%m/%d"
+    )
     day.delete_day = delete_day
     day.save()
 
 
 # pig_no = "77-77"
 # delete_day_set(pig_no)
+
 
 """ 全削除用 """
